requirements.txt
matplotlib
numpy
pandas
mplsoccer
statsmodels
itscalledsoccer
scipy
streamlit-ace
streamlit-metrics
altair
plotly